🚩 zv: `gvariant` no longer a default feature
Typical user doesn't need it and since cargo features are additive, this
ends up enabled when building from the repo (root-workspace) even if
zbus has it disabled, and zbus then has no way to track if this feature
is enabled or not.

Until now this has not been a problem since we do no have proper
gvariant support in zbus but that's about to change in a following
commit.