✨ zb: Allow messages to be in other encodings than `DBus`
This currently means `GVariant` if `gvariant` feature is enabled. We
translate this encoding to version 2 of the D-Bus spec, which was its
intended destiny.

Currently this will fail to work with either of the established bus
implementations so can only really be used on a p2p connection. However,
I plan to support this in busd.

This will break if there was a version 2 of the specification, that
doesn't use GVariant encoding but the chances of that happening are so
low that we really don't need to worry about it at all IMO.