⚗️  FAILED experiment: Make zbus::Address reference type
We've to decode the string in multiple properties involved here and the
input string is readonly so we've to allocate new strings anyway. We
could just keep the original string as is and decode on-demand but given
that most uses of Address will see these properties being read later,
that will only delay alloction. Moreover, it will be worse cause then
we'll have to allocate each time the properties in question are ready.

So I decided that we shouldn't bother with this, at least for zbus 4
cycle. Maybe we can have a look in the next breaking change.