🥅 zm: interface-generated proxy should use the same error types
If user specifies a Result type are return type of a method, the
generated proxy should use the same error types. We might need to add an
attribute to control this as this isn't always what the user would want.

We should do this because if we hardcode the error type to be
`zbus::Error`, the API will have to break if one decides to replace a
`proxy` use with this new ability of `interface` and they had their
`proxy` return a specific error type. This for example is the case for
`zbus::fdo` API, which we want to convert in a following commit.