♻️  zv: Use winnow for validation of Well-known names
Just like in the case for object paths in a previous commit in this
series, this change makes the parsing code **much** simpler and more
readable. However, unlike in case of the object path, it seems to make
parsing **much** faster (~43%). Although the header parsing benchmark
shows a 5% reduction in performance, we're talking a difference of around
10 nanoseconds here so even if the message header so not a huge deal to
worry about.

A "regression" here is the error message: it's now very generic. However,
it links to the relavent section in the specification so people will have
a way to figure it out.

Partially fixes #1070.