⚡️ zb: Sort interfaces by name in managed objects
For GetManagedObjects result, it would be more convenient if the
interfaces on each object path represented are sorted by interface
name instead of sent in random order.

This is not achieved by this change. While the interfaces are sorted in
the ManagedObjects structure, the marshalling procedure, which is
defined by the D-Bus spec, removes the order. The interfaces sent on the
D-Bus are in random order.

This change likely results in slightly more efficent processing when
building the temporary interfaces BTreeMap to create the ObjectManager
struct.

The interfaces, being ordered in Node's interfaces attribute, are inserted in
order into the temporary interfaces BTreeMap, so there is no cost involved
in shuffling the keys for sorting during insertion.

The space allocation cost is probably a toss up.

Signed-off-by: mulhern <amulhern@redhat.com>